# Use `map!` in `HashWithIndifferentAccess#fetch_values`